Rename json.py to _json.py
To make it explicit that it is a private implementation detail.
