+++ src/pip/_internal/metadata/_json.py
